refactor: rename deleteImage module to deleteImages
