+++ src/utils/deleteImages.js
